pyproject.toml
[tool.poetry]

name = "elections"
version = "0.0"
description = "Lookup registration and ballot data."

authors = ["Jace Browning <jacebrowning@gmail.com>"]

[tool.poetry.dependencies]

python = "^3.6"

django = "^2.0"
django-memoize = "^2.1"
django-grappelli = "^2.11"
django-filter = "*"

djangorestframework = "^3.8"
django-rest-swagger = "^2.2"

dj-database-url = "^0.5.0"
psycopg2-binary = "*"

django-redis-cache = "*"

markdown = "^2.6"
arrow = "*"

gunicorn = "^19.8"
whitenoise = "^3.3"

[tool.poetry.dev-dependencies]

isort = "^4.3"
black = "^18.3-alpha.0"

pylint = "^1.9"
pylint-django = "*"
mypy = "^0.610.0"

pytest = "^3.6"
pytest-django = "*"
pytest-describe = "*"
pytest-expecter = "*"
pytest-cov = "^2.5"
pytest-tldr = "*"
pytest-watch = "*"

factory_boy = "*"

django-livereload-server = "*"
